+++ src/Akkatecture/Aggregates/AggregateManagerSettings.cs
@@ -21,6 +21,7 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System;
 using Akka.Configuration;
 using Akkatecture.Configuration;
 
@@ -38,6 +39,7 @@ namespace Akkatecture.Aggregates
 
             HandleDeadLetters = aggregateManagerConfig.GetBoolean("handle-deadletters");
             AutoDispatchOnReceive = aggregateManagerConfig.GetBoolean("auto-dispatch-on-receive");
+
         }
     }
 }