+++ src/Akkatecture/Jobs/JobSchedulerFactory.cs
@@ -1,0 +1,10 @@
+using Akka.Actor;
+
+namespace Akkatecture.Jobs
+{
+    public class JobSchedulerFactory<TJob, TIdentity>
+        where TJob : IJob
+        where TIdentity : IJobId
+    {
+    }
+}