+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - IExecute interface as an alternative way to add command handlers to the aggregate.
+- JobScheduler and JobRunner actor types for scheduling jobs.
+- receive timeouts for aggregate roots.
+
+### Changed
+- AggregateRoot and AggregateSaga logging members are using Eventsourced.Log member.
 
 ## [0.4.6] - 2019-07-27